requirements_dev.txt
-r requirements.txt
pytest==6.2.5
pytest-responses==0.5.0
